+++ packages/jsii-pacmak/test/expected.jsii-calc-base/dotnet/Amazon.JSII.Tests.CalculatorPackageId.BasePackageId/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using Amazon.JSII.Runtime.Deputy;
 
-[assembly: JsiiAssembly("@scope/jsii-calc-base", "0.6.4", "scope-jsii-calc-base-0.6.4.tgz")]
+[assembly: JsiiAssembly("@scope/jsii-calc-base", "0.7.0", "scope-jsii-calc-base-0.7.0.tgz")]